Temporary migrate Chat Component from Chatview
