+++ frontend/src/views/ChatView.vue
@@ -18,41 +18,7 @@
         @select-thread="selectThread"
       />
     </div>
-    <div class="chat-container">
-      <ChatFrame>
-        <div v-for="(message, index) in messages" :key="index">
-          <!-- Response Container -->
-          <div class="message-container">
-            <section class="chat-response">
-              <MessageComponent 
-                :is-user="message.isUser"
-                :text="message.text" 
-                :typing="message.typing" 
-                :htmlContent="message.htmlContent"
-                :username="message.isUser ? displayName : 'FinBud Bot'"
-                :avatar-src="message.isUser ? userAvatar : botAvatar"
-                :sources="message.isUser ? [] : message.sources"
-                :videos="message.isUser ? [] : message.videos"
-                :relevantQuestions="message.isUser ? [] : message.relevantQuestions"
-                @question-click="handleQuestionClick"
-              />
-              <!-- <SourceComponent
-                v-if="sources.length"
-                :sources="sources"
-                class="source-component-card"
-              />
-              <FollowUpComponent
-                v-if="followUpQuestions.length"
-                :followUpQuestions="followUpQuestions"
-                @sendFollowUp="sendFollowUp"
-                class="followup-component-card"
-              /> -->
-            </section>
-          </div>
-        </div>
-      </ChatFrame>
-      <UserInput @send-message="sendMessage" @clear-message="clearMessage" />
-    </div>
+    <ChatComponent />
     <div 
       class="guidance-btn" 
       :class="{ 'is-guidance-visible': showGuidance }"
@@ -70,12 +36,10 @@
 <script>
 import authStore from "@/authStore";
 import axios from "axios";
-import ChatHeader from "../components/ChatHeader.vue";
-import ChatFrame from "../components/ChatFrame.vue";
-import MessageComponent from "../components/MessageComponent.vue";
+import ChatComponent from "@/components/ChatComponent.vue";
 // import SourceComponent from "@/components/SourceComponent.vue";
 // import FollowUpComponent from "@/components/FollowUpComponent.vue";
-import UserInput from "../components/UserInput.vue";
+
 import SideBar from "../components/SideBar.vue";
 import GuidanceModal from "../components/GuidanceModal.vue";
 import { gptServices } from "../services/gptServices.js";
@@ -85,10 +49,7 @@ export default {
   name: "ChatView",
   props: ["threadId"],
   components: {
-    ChatHeader,
-    ChatFrame,
-    MessageComponent,
-    UserInput,
+    ChatComponent,
     SideBar,
     GuidanceModal ,
     // SourceComponent,
@@ -96,26 +57,10 @@ export default {
   },
   data() {
     return {
-      newMessage: "",
-      messages: [],
-      sources: [],
-      followUpQuestions: [],
-      displayName: authStore.isAuthenticated
-        ? JSON.parse(localStorage.getItem("user")).identityData.displayName
-        : "User",
-      userAvatar: authStore.isAuthenticated
-        ? JSON.parse(localStorage.getItem("user")).identityData.profilePicture
-        : require("@/assets/anonymous.png"),
-      botAvatar: require("@/assets/botrmbg.png"),
       currentThread: {},
       threads: [],
       isSidebarVisible: false,
       showGuidance: false,
-      sources: [],
-      videos: [],
-      showVideos: false,
-      showSearchVideosButton: false,
-      relevantQuestions: [],
       overlayEnabled: false, //overlay to darken the chat screen when new window popsup
       newWindow: null, //new window to referrence to other
       windowCheckInterval: null,
@@ -137,9 +82,6 @@ export default {
     },
   },
   methods: {
-    clearMessage() {
-      this.newMessage = "";
-    },
     toggleSidebar() {
       this.isSidebarVisible = !this.isSidebarVisible;
     },
@@ -262,431 +204,7 @@ export default {
         console.error('Error on deleting thread or its associated chats:', err);
       }
     },
-    async scrollChatFrameToBottom() {
-      await new Promise(r => setTimeout(r, 200));
-      const chatFrame = document.querySelector(".chat-frame");
-      chatFrame.scrollTo({
-        top: chatFrame.scrollHeight,
-        behavior: 'smooth'
-      });
-    },
-    async sendMessage(newMessage) {
-      const userMessage = newMessage.trim();
-      //ONLY EXECUTE COMMAND/SHOW PROMPT IF THERE IS SOME MESSAGES IN THE USER INPUT
-      if (userMessage.length != 0) {
-        this.messages.push({
-          text: userMessage,
-          isUser: true,
-          typing: true,
-          timestamp: new Date().toLocaleTimeString(),
-        });
-        const answers = [];
-        let newSources = [];
-        let newVideos = [];
-        let newRelevantQuestions = [];
-
-        // Define additional parameters
-        const returnSources = true;
-        const textChunkSize = 800;
-        const textChunkOverlap = 200;
-        const numberOfSimilarityResults = 2;
-        const numberOfPagesToScan = 4;
-
-
-        // HANDLE DEFINE
-        if (userMessage.toLowerCase().includes("define")) {
-          try {
-            const term = userMessage
-              .substring(
-                userMessage.toLowerCase().indexOf("define") + "define".length
-              )
-              .trim();
-            const prompt = `Explain ${term} to me as if I'm 15.`;
-            const gptResponse = await gptServices(prompt);
-            answers.push(gptResponse);
-          } catch (err) {
-            console.error("Error in define message:", err);
-          }
-        }
-        // HANDLE BUY (7)
-        if (userMessage.toLowerCase().includes("buy")) {
-          try {
-            const buyRegex = /#buy\s+([A-Z]+)\s+(\d+)/i;
-            const match = userMessage.match(buyRegex);
-            if (match) {
-              const stockSymbol = match[1].toUpperCase();
-              const quantity = parseInt(match[2], 10);
-              if (stockSymbol && !isNaN(quantity)) {
-                const url = this.$router.resolve({
-                  path: "/stock-simulator",
-                  query: { symbol: stockSymbol, quantity },
-                }).href;
-                this.openNewWindow(url);
-              } else {
-                this.addTypingResponse(
-                  "Invalid stock symbol or quantity",
-                  false
-                );
-              }
-            } else {
-              this.addTypingResponse("Invalid buy command format", false);
-            }
-          } catch (err) {
-            console.error("Error in buy message:", err);
-          }
-        }
-        // HANDLE SELL (8)
-        else if (userMessage.toLowerCase().includes("sell")) {
-          try {
-            const sellRegex = /#sell\s+([A-Z]+)\s+(\d+)/i;
-            const match = userMessage.match(sellRegex);
-            if (match) {
-              const stockSymbol = match[1].toUpperCase();
-              const quantity = parseInt(match[2], 10);
-              if (stockSymbol && !isNaN(quantity)) {
-                const url = this.$router.resolve({
-                  path: "/stock-simulator",
-                  query: { symbol: stockSymbol, quantity: -quantity },
-                }).href;
-                this.openNewWindow(url);
-              } else {
-                this.addTypingResponse(
-                  "Invalid stock symbol or quantity",
-                  false
-                );
-              }
-            } else {
-              this.addTypingResponse("Invalid sell command format", false);
-            }
-          } catch (err) {
-            console.error("Error in sell message:", err);
-          }
-        }
-        // HANDLE ADD TRANSACTION (5)
-        else if (userMessage.toLowerCase().includes("#add")) {
-          try {
-            const match = userMessage.match(/#add\s+([\w\s]+)\s+(\d+)/i);
-            if (match) {
-              const description = match[1].trim();
-              const amount = parseInt(match[2], 10);
-              const balance = await this.calculateNewBalance(amount);
-              await this.addTransaction(description, amount, balance);
-              answers.push(
-                `Transaction added: ${description}, $${amount}. New balance: $${balance}.`
-              );
-              this.openNewWindow("/goal");
-            } else {
-              answers.push(
-                "Please specify the description and amount you want to add."
-              );
-            }
-          } catch (err) {
-            console.error("Error in add transaction:", err);
-          }
-        }
-        // HANDLE SPEND TRANSACTION (6)
-        else if (userMessage.toLowerCase().includes("#spend")) {
-          try {
-            const match = userMessage.match(/#spend\s+([\w\s]+)\s+(\d+)/i);
-            if (match) {
-              const description = match[1].trim();
-              const amount = -parseInt(match[2], 10);
-              const balance = await this.calculateNewBalance(amount);
-              await this.addTransaction(description, amount, balance);
-              answers.push(
-                `Transaction spent: ${description}, $${Math.abs(
-                  amount
-                )}. New balance: $${balance}.`
-              );
-              this.openNewWindow("/goal");
-            } else {
-              answers.push(
-                "Please specify the description and amount you want to spend."
-              );
-            }
-          } catch (err) {
-            console.error("Error in spend transaction:", err);
-          }
-        }
-        // HANDLE STOCK
-        else if (this.extractStockCode(userMessage)) {
-          try {
-            const stockCode = this.extractStockCode(userMessage)[0];
-            const stockResponse = await axios.get(
-              `https://www.alphavantage.co/query?function=GLOBAL_QUOTE&symbol=${stockCode}&apikey=${process.env.VUE_APP_ALPHA_VANTAGE_API_KEY}`
-            );
-            const stockData = stockResponse.data;
-            const price = stockData["Global Quote"]["05. price"];
-            const timeStamp = new Date().toLocaleTimeString();
-            let alphavantageResponse = `The current price of ${stockCode} stock is $${price}, as of ${timeStamp}.`;
-            answers.push(alphavantageResponse);
-            //chatgpt api
-            const prompt = `Generate a detailed analysis of ${stockCode} which currently trades at $${price}.`;
-            const gptResponse = await gptServices(prompt);
-            answers.push(gptResponse);
-          } catch (err) {
-            console.error("Error in stock message:", err);
-          }
-        }
-        // RETURNS CRYPTO TABLE (3)
-        else if (userMessage.toLowerCase().includes("#crypto")) {
-          //FETCHING COIN DATA
-
-          let coinData = [];
-          try {
-            const res = await axios.get(
-              "https://api.coinranking.com/v2/coins?timePeriod=7d",
-              {
-                headers: {
-                  "x-access-token": process.env.VUE_APP_COINRANKING_KEY,
-                },
-              }
-            );
-            coinData = res.data.data.coins;
-          } catch (err) {
-            console.error("Failed to fetch cr quotes:", err);
-          }
-          let tableTemplate = `
-        <div style="font-weight: 900; font-size: 30px"> Top 5 Ranking Coins </div>
-        <table>
-          <thead>
-            <tr>
-              <th>Name</th>
-              <th>Rank</th>
-              <th>Tier</th>
-              <th>Price</th>
-              <th>Symbol</th>
-              <th>Change</th>
-            </tr>
-          </thead>
-          <tbody id="tableBody" class="table-body">`;
-          coinData.slice(0, 5).map((item) => {
-            tableTemplate += `
-            <tr>
-              <td><img style="width: 50px; aspect-ratio: 1;" src=${
-                item.iconUrl
-              } alt=${item.name}>${item.name}</td>
-              <td>${item.rank}</td>
-              <td>${item.tier}</td>
-              <td>${parseFloat(item.price).toFixed(2)}$</td>
-              <td>${item.symbol}</td>
-              <td>${item.change}</td>
-            </tr>
-          `;
-          });
-          tableTemplate += `</tbody></table>`;
-          this.messages.push({
-            text: ``,
-            htmlContent: tableTemplate,
-            isUser: false,
-            typing: true,
-            timestamp: new Date().toLocaleTimeString(),
-          });
-        }
-        // RETURNS REALESTATE TABLE
-        else if (userMessage.toLowerCase().includes("#realestate")) {
-          let userInputToken = userMessage.toLowerCase().split(/\s+/);
-          let searchLocation;
-          if (userInputToken.length > 1) {
-            userInputToken = userInputToken.slice(1, userInputToken.length);
-            searchLocation = userInputToken.join(" ");
-          } else {
-            searchLocation = "san jose";
-          }
-          let propertiesData = [];
-          const API_KEY = process.env.VUE_APP_REAL_ESTATE_KEY;
-          const BASE_URL =
-            "https://zillow-com1.p.rapidapi.com/propertyExtendedSearch";
-          try {
-            const response = await axios.get(BASE_URL, {
-              params: { location: searchLocation },
-              headers: {
-                "X-RapidAPI-Key": API_KEY,
-                "X-RapidAPI-Host": "zillow-com1.p.rapidapi.com",
-              },
-            });
-            propertiesData = response.data.props;
-          } catch (err) {
-            console.error("Error fetching property data:", err);
-          }
-          let tableTemplate = `
-        <div style="font-weight: 900; font-size: 30px"> Listing of 5 Properties in ${searchLocation} </div>
-        <table>
-          <thead>
-            <tr>
-              <th>Image</th>
-              <th>Type</th>
-              <th>Address</th>
-              <th>Price</th>
-              <th>Status</th>
-            </tr>
-          </thead>
-          <tbody id="tableBody" class="table-body">`;
-          propertiesData.slice(0, 5).map((item) => {
-            tableTemplate += `
-            <tr>
-              <td><img style="width: 50px; aspect-ratio: 1;" src=${item.imgSrc} alt="propertyImage"></td>
-              <td>${item.propertyType}</td>
-              <td>${item.address}</td>
-              <td>${item.price}$</td>
-              <td>${item.listingStatus}</td>
-            </tr>`;
-          });
-          tableTemplate += `</tbody></table>`;
-          this.messages.push({
-            text: ``,
-            htmlContent: tableTemplate,
-            isUser: false,
-            typing: true,
-            timestamp: new Date().toLocaleTimeString(),
-          });
-        }
-        // HANDLE SEARCH
-        else if (userMessage.toLowerCase().includes("#search")){
-          //Search for sources, videos, and relevant questions
-          const searchResults = await getSources(userMessage);
-          newSources = searchResults;
-          newVideos = await getVideos(userMessage);
-          newRelevantQuestions = await getRelevantQuestions(searchResults);
-          //Normal GTP response
-          const gptResponse = await gptServices(userMessage);
-          answers.push(gptResponse);
-          // this.showSearchVideosButton = true; ??????
-        }
-        // HANDLE GENERAL
-        else {
-          try {
-            const prompt = userMessage;
-            // const gptResponse = await axios.post(
-            //   `${process.env.VUE_APP_DEPLOY_URL}/query`,
-            //   {
-            //     prompt,
-            //     returnSources,
-            //     textChunkSize,
-            //     textChunkOverlap,
-            //     numberOfSimilarityResults,
-            //     numberOfPagesToScan,
-            //   }
-            // );
-            const gptResponse =  await gptServices(prompt);
-            answers.push(gptResponse);
-            // answers.push(gptResponse.data.answer);
-            // this.sources = gptResponse.data.sources || [];
-            // this.followUpQuestions = gptResponse.data.followUpQuestions || [];
-          } catch (err) {
-            console.error("Error in general message:", err);
-          }
-        }
-        answers.forEach((answer) => {
-          this.addTypingResponse(answer, false, newSources, newVideos, newRelevantQuestions);
-        });
-        //save chat to backend
-        if (authStore.isAuthenticated) {
-          try {
-            const chatApi = `${process.env.VUE_APP_DEPLOY_URL}/chats`;
-            const reqBody = {
-            prompt: userMessage,
-            response: answers,
-            sources: newSources,
-            videos: newVideos,
-            followUpQuestions: newRelevantQuestions,
-            threadId: this.currentThread.id,
-          };
-            const chat = await axios.post(chatApi, reqBody);
-          } catch (err) {
-            console.error("Error on saving chat:", err);
-          }
-        }
-        this.scrollChatFrameToBottom();
-      }
-    },
-    async handleAddTransaction(userMessage) {
-      const match = userMessage.match(/#add\s+([\w\s]+)\s+(\d+)/i);
-      if (match) {
-        const description = match[1].trim();
-        const amount = parseInt(match[2], 10);
-        const balance = await this.calculateNewBalance(amount);
-        await this.addTransaction(description, amount, balance);
-        return [
-          `Transaction added: ${description}, $${amount}. New balance: $${balance}.`,
-        ];
-      } else {
-        return ["Please specify the description and amount you want to add."];
-      }
-    },
-    async handleSpendTransaction(userMessage) {
-      const match = userMessage.match(/#spend\s+([\w\s]+)\s+(\d+)/i);
-      if (match) {
-        const description = match[1].trim();
-        const amount = -parseInt(match[2], 10);
-        const balance = await this.calculateNewBalance(amount);
-        await this.addTransaction(description, amount, balance);
-        return [
-          `Transaction spent: ${description}, $${Math.abs(
-            amount
-          )}. New balance: $${balance}.`,
-        ];
-      } else {
-        return ["Please specify the description and amount you want to spend."];
-      }
-    },
-    async addTransaction(description, amount, balance) {
-      try {
-        const response = await axios.post(`${process.env.VUE_APP_DEPLOY_URL}/transactions`, {
-          description,
-          amount,
-          balance,
-          date: new Date().toISOString(),
-          userId: localStorage.getItem('token')
-        });
-      } catch (err) {
-        console.error('Error adding transaction:', err);
-        this.addTypingResponse('Error adding transaction.', false);
-      }
-    },
-    async calculateNewBalance(amount) {
-      try {
-        const userId = localStorage.getItem("token");
-        const response = await axios.get(
-          `${process.env.VUE_APP_DEPLOY_URL}/transactions/u/${userId}`
-        );
-        const transactions = response.data;
-        const currentBalance = transactions.reduce(
-          (acc, transaction) => acc + transaction.amount,
-          0
-        );
-        return currentBalance + amount;
-      } catch (err) {
-        console.error('Error calculating new balance:', err);
-        throw error;
-      }
-    },
-    extractStockCode(message) {
-      const pattern = /\b[A-Z]{3,5}\b/g;
-      const matches = message.match(pattern);
-      return matches;
-    },
-    handleQuestionClick(question) {
-      const searchQuery = `#search ${question}`;
-      this.sendMessage(searchQuery);
-    },
-    addTypingResponse(text, isUser, sources = [], videos = [], relevantQuestions = []) {
-      const typingMessage = {
-        text: text,
-        isUser: isUser,
-        typing: true,
-        timestamp: new Date().toLocaleTimeString(),
-        username: isUser ? "You" : "FinBud Bot",
-        sources: sources,
-        videos: videos,
-        relevantQuestions: relevantQuestions
-      };
-      this.messages.push(typingMessage);
-      setTimeout(() => {
-        typingMessage.typing = false;
-        this.$forceUpdate();
-      }, 1000);
-    },
+   
     //USED IN BUY/SELL/ADD/SPEND/(QUIZ?)
     openNewWindow(url) {
       const screenWidth = window.screen.width;
@@ -727,14 +245,6 @@ export default {
       window.removeEventListener("click", this.closeOnClickOutside);
       this.overlayEnabled = false;
       this.newWindow = null;
-    },
-    async scrollChatFrameToBottom() {
-      await new Promise((r) => setTimeout(r, 200));
-      const chatFrame = document.querySelector(".chat-frame");
-      chatFrame.scrollTo({
-        top: chatFrame.scrollHeight,
-        behavior: "smooth", // Smooth scrolling effect
-      });
     },
   },
   async mounted() {
